Automatically generate voting categories (axises) matching field names (without the 'field_' prefix).
Patch at http://drupal.org/node/1246656#comment-4854242.

As of this commit, there is no more configuration page for Fivestar.
The only configuration left was Voting Categories, which is set
automatically now.

Remove hook_menu(). In addition to the config page, there was an old
color picker page lurking here. This page is no longer exists. It was just cruft.