AN-27704, Implement delete hooks for nodes and fields that get deleted.
This commit improves the situation for Drupal Gardens 1.07 release on August 17, 2011,
but does not solve it.

This patch only really works for nodes and is only a partial fix.

Implementing hook_field_instance_delete() is a step in the right direction. But
both EntityFieldQuery and Voting API don't really handle bundles, which makes it difficult
to support all entities that might potentially have fivestar fields. (For more detail,
see @todo notes in comments.)